Delete RooCintUtils.h and RooCintUtils.cxx
I know this is not a small change, so any comments are welcome!

Background:
RooCintUtils was providing a wrapper between Cint and roofit. This implementation is very old, and is
using Cint iteration interface a lot. Recently we were trying to kill
RooCintUtils code where its using iteration and to change them to new
Cling interface in order to improve the performance.

RooCintUtils is an utility class, so this is internal namespace which is only used inside roofit.
Since there are just 3 methods left, this RooCintUtils is no longer
worth spending a file. Let's delete this and migrate methods to
RooFactoryWSTool.
