+++ roofit/roofitcore/Module.mk
@@ -77,7 +77,7 @@ ROOFITCOREH4   := RooConstraintSum.h RooRecursiveFraction.h RooDataWeightedAvera
                   RooSimWSTool.h RooFracRemainder.h RooAbsCachedReal.h \
                   RooAbsSelfCachedReal.h RooCachedReal.h RooNumCdf.h RooChangeTracker.h \
                   RooNumRunningInt.h RooHistFunc.h RooExpensiveObjectCache.h \
-                  RooBinningCategory.h RooCintUtils.h RooFactoryWSTool.h RooTFoamBinding.h RooFunctor.h	\
+                  RooBinningCategory.h RooFactoryWSTool.h RooTFoamBinding.h RooFunctor.h	\
                   RooDerivative.h RooGenFunction.h RooMultiGenFunction.h RooAdaptiveIntegratorND.h \
                   RooAbsNumGenerator.h RooFoamGenerator.h RooNumGenConfig.h RooNumGenFactory.h \
                   RooMultiVarGaussian.h RooXYChi2Var.h RooAbsDataStore.h RooTreeDataStore.h RooTreeData.h \
